+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added basic support for the STAC API extensions Transactions (for Items) and Collection Transactions
+  - Adds three new config options: `transactions`, `transactionsRequireLogin` and `transactionsRequirePreflight`
+
 ## [5.0.0-rc.2] - 2026-06-23
 
 ### Added